test(overlays): migrate
